Add internal/argon2 package
This adds an internal package for deriving keys from passphrases
using the Argon2i algorithm, and will be used for passphrase support.

There are 2 APIs:
- Benchmark determines the cost parameters for the algorithm based
on the supplied target duration, maximum memory cost and key length.
It is based on the algorithm used by cryptsetup, which prioritizes
memory cost over time cost.
- Key is just a wrapper around the argon2.Key API provided by the
golang.org/x/crypto/argon2 package, but using the CostParams type
returned by Benchmark.

Benchmark has to run the garbage collector between each measurement
else it will rapidly consume all of the available memory. For this
reason, it should be executed in a short-lived helper process that
will live outside of the internal/argon2 package. It definitely
shouldn't be executed inside a long-lived system process.

Key doesn't run the garbage collector but does temporarily consume
a lot of memory depending on the supplied parameters. It's probably
best that this API isn't used inside a long-lived system process
either.