preinstall: Add checks for TPM2 device
This adds checks that we have a TPM2 device, that it is enabled, that it
is not in DA lockout mode, that it is not already owned by having an
authorization value for any of the 3 hierarchies we can use, that it is
a PC-Client device, and that it has enough spare NV counter indexes to
supprt PCR policy revocation. If the checks succeed, it returns a
tpm2.TPMContext that will be consumed by RunChecks for other tests, and
which will have to be closed later on.